+++ niftilib/nifti1_tool.c
@@ -829,7 +829,7 @@ add_int(int_list * ilist, int val)
   if (ilist->len == 0)
     ilist->list = NULL; /* just to be safe */
   ilist->len++;
-  ilist->list = (int *)realloc(ilist->list, ilist->len * sizeof(int));
+  ilist->list = (int *)realloc(ilist->list, (size_t)ilist->len * sizeof(int));
   if (!ilist->list)
   {
     fprintf(stderr, "** failed to alloc %d (int *) elements\n", ilist->len);
@@ -853,7 +853,7 @@ add_string(str_list * slist, const char * str)
   if (slist->len == 0)
     slist->list = NULL; /* just to be safe */
   slist->len++;
-  slist->list = (const char **)realloc(slist->list, slist->len * sizeof(char *));
+  slist->list = (const char **)realloc(slist->list, (size_t)slist->len * sizeof(char *));
   if (!slist->list)
   {
     fprintf(stderr, "** failed to alloc %d (char *) elements\n", slist->len);
@@ -1925,7 +1925,7 @@ read_file_text(const char * filename, int * length)
 
   /* allocate the bytes, and fill them with the file contents */
 
-  text = (char *)malloc(len * sizeof(char));
+  text = (char *)malloc((size_t)len * sizeof(char));
   if (!text)
   {
     fprintf(stderr, "** RFT: failed to allocate %d bytes\n", len);
@@ -2138,7 +2138,7 @@ remove_ext_list(nifti_image * nim, const char ** elist, int len)
   if (g_debug > 2)
     fprintf(stderr, "+d removing %d exts from '%s'\n", len, nim->fname);
 
-  if (!(marks = (int *)calloc(nim->num_ext, sizeof(int))))
+  if (!(marks = (int *)calloc((size_t)(nim->num_ext), (size_t)sizeof(int))))
   {
     fprintf(stderr, "** failed to alloc %d marks\n", nim->num_ext);
     return -1;
@@ -4434,8 +4434,8 @@ nt_read_bricks(nt_opts * opts, const char * fname, int len, int * list, nifti_br
 
   /* now populate NBL (can be based only on len and nim) */
   NBL->nbricks = len;
-  NBL->bsize = (size_t)nim->nbyper * nim->nx * nim->ny * nim->nz;
-  NBL->bricks = (void **)calloc(NBL->nbricks, sizeof(void *));
+  NBL->bsize = (size_t)nim->nbyper * (size_t)nim->nx * (size_t)nim->ny * (size_t)nim->nz;
+  NBL->bricks = (void **)calloc((size_t)(NBL->nbricks), (size_t)(sizeof(void *)));
   if (!NBL->bricks)
   {
     fprintf(stderr, "** NRB: failed to alloc %d pointers\n", NBL->nbricks);
